Initial Load of Files
This is the initial load of all the local files for the project.

Basic functionality of the original WTFEngine has been replicated. Still
missing extra effects, such as animation and ability to use a Google
Spreadsheet.

The tests directory contains test code for both the coffee / javascript
and a mostly complete implementation of the idea in Python.