Added capability to read nested objects
Added the capability to read nested objects in the data file. This
allows for better visual organization and is the next step for the
planned feature of being able to customize the output by dynamically
turning on and off certain lists in the data.

Added more entries to the character data set.

Started data set for magic items, also usefull to test ability to
correctly use any data file.