Added animation effects, edit README
Finally added the animation effects upon text generation and got them to work correctly. Currently implemented in CSS with a bit of jQuery to correctly sync them. An alternate implemenation using only jQuery is in the file as comments.

Also added more entries to the database lists, and rewrote the README.