Improvements to Navbar; Checkboxes
Minor edits and improvements to the navbar mainly consisting of tweaks
to the CSS. Wrote the navbar javascript in coffeescript.

Added CSS3 checkboxes and implemented them in a list in the navbar. They
currently do nothing except look decent and provide a place to test
code.

Also added a few more entries to the generation lists.