Added About Page, Started Dynamic Lists
Quite a few changes. Added an About page along with a place holder for a Suggestions page, and made the links in the navbar work. Added CSS to style the about page.

Implemented code to handle dynamically changing the generator lists. The code now builds a config list containing true and false values from the Objects inside the generator data. The build_list function then inlcudes or excludes parts of the generator data based on the values in the config list.

Added some test code to change the config list and regenerate the data list through the checkboxes in the navbar. Had to change data_list to a global variable via window but it now gives the correct end result.