Finished dynamic generation lists
Implemented code to create a checkbox for each sub-object in the
generation list and add it to the navbar. The checkboxes now enable /
disable that particular sub-list for the generator.

The CSS was tweaked slightly to keep formatting. The checkboxes were
also coded to start checked by default.

A couple random typos were corrected and various outdated comments were
deleted.